% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_tests.R
\name{capture_log1}
\alias{capture_log1}
\title{Capture message, warnings and errors from a R function}
\usage{
capture_log1(f)
}
\arguments{
\item{f}{Function to gather the logs from.}
}
\value{
List of logs.
}
\description{
Capture message, warnings and errors from a R function
}
\examples{
#' @examples
#' \dontrun{
#' test_message <- function(a) {
#'   message("this is test from function message.")
#'   return(a)
#' }
#'
#' test_warning <- function(a) {
#'   warning("this is test from function warning.")
#'   return(a)
#' }
#'
#' test_error <- function(a) {
#'   stop("this is test from function error.")
#'   return(a)
#' }
#'
#' capture_log1(test_message)(1)
#'
#' capture_log1(test_warning)(1)
#'
#' capture_log1(test_error)(1)
#' }
}
\references{
\url{https://www.r-bloggers.com/2020/10/capture-message-warnings-and-errors-from-a-r-function/#google_vignette}
}
