% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{clustmeth_identifier_number}
\alias{clustmeth_identifier_number}
\title{Create a single identifier number}
\usage{
clustmeth_identifier_number(method_name, path)
}
\arguments{
\item{method_name}{Name of the clustering method.}

\item{path}{Directory path.}
}
\value{
\code{identifier_number}: Container identifier.
}
\description{
Function to create a single identifier number.
}
