% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_param.R
\name{copy_param}
\alias{copy_param}
\title{Function copying the description and the input path from files to the
PARAM_GLOBAL}
\usage{
copy_param(path, path_out)
}
\arguments{
\item{path}{path of the parameter folder.}

\item{path_out}{directory to store the output folder.}
}
\description{
Writes the description and the input path from text files to the
PARAM_GLOBAL.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
