% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr_stat_fun.R
\name{descr_stat_fun}
\alias{descr_stat_fun}
\title{Function to generate descriptive statistics}
\usage{
descr_stat_fun(descrstat_dpath, lvalues)
}
\arguments{
\item{descrstat_dpath}{Directory to store the tex files containing the
descriptive statistics.}

\item{lvalues}{List of variables and values to be analysed. The list must
have variable names.}
}
\value{
Latex tables.
}
\description{
Generate descriptive statistics as the mean, the median, among others.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
