% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{draw_flow}
\alias{draw_flow}
\title{Draw data flow}
\usage{
draw_flow(DF)
}
\arguments{
\item{DF}{Data frame to describe in the data flow.}
}
\value{
Graph in the RStudio Viewer.
}
\description{
Draw data flow.
}
\author{
\href{mailto:christoph@christophsax.com}{Christoph Sax}
}
