% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{fun_ggplot_hist2}
\alias{fun_ggplot_hist2}
\title{Function to make histograms using clusters distribution}
\usage{
fun_ggplot_hist2(dta)
}
\arguments{
\item{dta}{Filtered data frame for each desired population sample}
}
\description{
Function to make histograms using clusters distribution
}
