% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plot.R
\name{function_kamplot}
\alias{function_kamplot}
\title{Draw scatterplot using clusters distribution}
\usage{
function_kamplot(dta)
}
\arguments{
\item{dta}{Filtered data frame for each desired population sample.}
}
\description{
Draw scatterplot using clusters distribution.
}
