% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_demo_data.R
\name{gen_demo_data}
\alias{gen_demo_data}
\title{Generate random demo data}
\usage{
gen_demo_data(path = tempdir(), method_name = "kamila", data_size = 1000)
}
\arguments{
\item{path}{Path where to store the demo data. Default: \code{tempdir()}.}

\item{method_name}{Name of the clustering method.}

\item{data_size}{Size of the simulated data frame. Default: 1000 rows.}
}
\value{
The path of the csv file.
}
\description{
Generate random demo data and write into a csv file
}
\examples{
gen_demo_data()
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
