% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{invalid_utf8}
\alias{invalid_utf8}
\title{Search for invalid lines in UTF-8}
\usage{
invalid_utf8(x)
}
\arguments{
\item{x}{Object to convert into UTF-8.}
}
\description{
Search for invalid lines in UTF-8.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
