% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_calc_kamila.R
\name{mod_calc_kamila}
\alias{mod_calc_kamila}
\title{Splitting the initial dataset into kstar clusters.}
\usage{
mod_calc_kamila(
  PARAM_KAMILA,
  CONT_DF,
  CATEG_DF,
  FULL_CONT_DF,
  FULL_CATEG_DF,
  KM_RES,
  list = NULL
)
}
\arguments{
\item{PARAM_KAMILA}{data frame with all needed parameters for the Kamila method,
from which the following parameters are used:
\itemize{
\item \code{numinit}: The number of initializations used.
\item \code{maxiter}: The maximum number of iterations in each run.
\item \code{param_kstar}: Best number of clusters estimated in the module
\code{\link{mod_kstar}}.
}}

\item{CONT_DF}{subset of the register of rents containing all the continuous
variables except for the outcome variables aadr and monthly_pension.}

\item{CATEG_DF}{subset of the register of rents containing all categorical
variables as factors except for the nominal variables marital_stat and benef_type.}

\item{FULL_CONT_DF}{data frame containing the continuous variables used for the
estimation plus the outcome variables aadr and monthly_pension.}

\item{FULL_CATEG_DF}{data frame containing the categorical variables used for
the estimation plus the nominal variables marital_stat and benef_type.}

\item{KM_RES}{tibble contains the kamila results of the training set.}

\item{list}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidy data frames:
\itemize{
\item \code{PLOTDATKAM} data frame containing the clusters factor and the other
variables.
\item \code{KM_RES_FINAL} data frame containing the resulting parameters of the
clustering.
\item \code{CONTVARS} data frame containing the continuous standardised variables.
\item \code{FULL_CONT_DF} data frame containing the continuous variables used for the
estimation.
\item \code{FULL_CATEG_DF} data frame containing the categorical variables used for
the estimation.
}
}
\description{
Splitting the initial dataset into kstar clusters by using the
parameter kstar determined in the module \code{\link{mod_kstar}}.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
