% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{mod_init}
\alias{mod_init}
\title{Check whether all inputs are tidy tibbles}
\usage{
mod_init(mod.function = NULL)
}
\arguments{
\item{mod.function}{Name of the module.}
}
\value{
Invisible.
}
\description{
Check whether all inputs are tidy tibbles.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
