% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_inp_kamila.R
\name{mod_inp_kamila}
\alias{mod_inp_kamila}
\title{Function to read the data}
\usage{
mod_inp_kamila(path, list = NULL, method_name = "kamila")
}
\arguments{
\item{path}{Path of the set of parameters set to a particular output.}

\item{list}{Path where the outputs are saved.}

\item{method_name}{Name of the clustering method.}
}
\value{
tl_inp Tidylist of inputs.
}
\description{
Function to read the data.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
