% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_input_ind_yearly_rr.R
\name{mod_input_ind_yearly_rr}
\alias{mod_input_ind_yearly_rr}
\title{Read of the pension register}
\usage{
mod_input_ind_yearly_rr(PARAM_INPUTS, list = NULL)
}
\arguments{
\item{PARAM_INPUTS}{a data frame containing the data paths.}

\item{list}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidy data frames:
\itemize{
\item \code{IND_YEARLY_RR} : contains all the beneficiaries of the OASI
}
}
\description{
Read the pension register and save it into an tidy data frame.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
