% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_kstar.R
\name{mod_kstar}
\alias{mod_kstar}
\title{Estimation of the best number of clusters using the Kamila algorithm.}
\usage{
mod_kstar(PARAM_KAMILA, CATEG_DF_TS, CONT_DF_TS, list = NULL)
}
\arguments{
\item{PARAM_KAMILA}{data frame with all needed parameters for the Kamila method,
from which the following parameters are used:
\itemize{
\item \code{numberofclusters}: The number of clusters returned by the algorithm, i.e.
sequence indicating the number of clusters which should be investigated to
extract the optimal number of clusters.
\item \code{numinit}: The number of initializations used.
\item \code{maxiter}: The maximum number of iterations in each run.
\item \code{calcnumclust}: Character: Method for selecting the number of clusters. Setting
calcNumClust to ’ps’ uses the prediction strength method of
Tibshirani &Walther (J. of Comp. and Graphical Stats. 14(3), 2005).
\item \code{pred_threshold}: Threshold fixed to 0.8 for well separated clusters (i.e.
not overlapping).
}}

\item{CATEG_DF_TS}{Training set of the register of rents containing all categorical
variables as factors.}

\item{CONT_DF_TS}{Training set of the register of rents containing all the continuous
variables.}

\item{list}{List of input data frames.}
}
\value{
a tidylist containing the following tidy data frames:
\itemize{
\item \code{KM_RES} data frame containing the results of the clustering.
\item \code{PARAM_KAMILA} data frame with the updated kstar parameter.
}
}
\description{
Estimation of best number of clusters using the Kamila algorithm
on the training set.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
