% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_log.R
\name{mod_log}
\alias{mod_log}
\title{Write the version of the packages}
\usage{
mod_log(list = NULL)
}
\arguments{
\item{list}{List of input data frames.}
}
\value{
\code{LOG} Tibble with information about time and packages version.
}
\description{
Write the version of the rrclust and dplyr packages and the time of the
output production.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
