% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_out.R
\name{mod_out_kamila}
\alias{mod_out_kamila}
\title{Output Kamila}
\usage{
mod_out_kamila(path, path_out, tl_inp_kamila, tl_out_kamila)
}
\arguments{
\item{path}{vector containing container path.}

\item{path_out}{vector containing output path.}

\item{tl_inp_kamila}{list containing input data Kamila.}

\item{tl_out_kamila}{list containing output data Kamila.}
}
\value{
path_out_identifier path to output container
}
\description{
Saves the Kamila output
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
