% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_prepa_rr.R
\name{mod_prepa_rr}
\alias{mod_prepa_rr}
\title{Preparation of the register of rents data.}
\usage{
mod_prepa_rr(IND_YEARLY_RR, list = NULL)
}
\arguments{
\item{IND_YEARLY_RR}{a data frame containing the data of the register of rents
subsetted for one year only.}

\item{list}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidy data frames:
\itemize{
\item \code{RR_OASI} : contains all the beneficiaries of the OASI.
}
}
\description{
Prepares the variables of the register of rents.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
