% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{mod_return}
\alias{mod_return}
\title{mod_return() for the modules}
\usage{
mod_return(...)
}
\arguments{
\item{...}{any tidylist containing tibble or simply tibble.}
}
\description{
Checks whether all the ouptus are tibbles.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
