% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tsvs.R
\name{mod_tsvs}
\alias{mod_tsvs}
\title{Splitting the data into a Training and a Validation sets}
\usage{
mod_tsvs(RR_OASI, PARAM_GLOBAL, list = NULL)
}
\arguments{
\item{RR_OASI}{a data frame containing the all the data.}

\item{PARAM_GLOBAL}{a data frame containing the parameters. We use the following:
\itemize{
\item \code{pct_sample_ts}: percentage of observations which build the training set.
}}

\item{list}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidy data frames:
\itemize{
\item \code{RR_OASI_TS} : Training set of categorical data.
\item \code{RR_OASI_VS}: Validation set of categorical data.
}
}
\description{
Splits the data into a training and a validation sets, given
the too large number of observations, in order to determine the best number of
clusters for the KAMILA algorithm.
}
\references{
\href{https://www.geeksforgeeks.org/the-validation-set-approach-in-r-programming/}{www.geeksforgeeks.org}
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
