% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_run.R
\name{multi_run}
\alias{multi_run}
\title{Write a message and execute the function applied on one or more containers}
\usage{
multi_run(path, fun, ...)
}
\arguments{
\item{path}{Path of the parameters container.}

\item{fun}{Function to run for each of the given paths.}

\item{...}{Any other arguments.}
}
\description{
Write a message and execute the function applied on one or more containers.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
