% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_replace.R
\name{param_replace}
\alias{param_replace}
\title{Build the tibble PARAM_ALL}
\usage{
param_replace(tl, ..., param_replace = NULL, update_param_all = TRUE)
}
\arguments{
\item{tl}{tidylist}

\item{...}{any other arguments}

\item{param_replace}{parameter which can be used to test some other parameters.}

\item{update_param_all}{if TRUE, update all parameters}
}
\description{
Build the tibble PARAM_ALL
}
\author{
\href{mailto:christoph@christophsax.com}{Christoph Sax}
}
