% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_resolve_path_out.R
\name{path_out_resolve}
\alias{path_out_resolve}
\title{Function resolving the path_out}
\usage{
path_out_resolve(path_out, identifier_number)
}
\arguments{
\item{path_out}{directory to store the output folder.}

\item{identifier_number}{unique identifier number to name the output folder.}
}
\value{
path_out directory to store the output
}
\description{
Writes the path_out and the unique identifier_number of the output.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
