% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{path_param}
\alias{path_param}
\title{Function to create a the path of the parameters}
\usage{
path_param(path)
}
\arguments{
\item{path}{directory path}
}
\description{
Function to create a the path of the parameters
}
