% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Read and write the inputs for the calculation in a tidylist}
\usage{
prepare_input(path, path_out = file.path(dirname(path)), overwrite = FALSE)
}
\arguments{
\item{path}{Path to container.}

\item{path_out}{Path to output container.}

\item{overwrite}{if TRUE, the output should be overwritten.}
}
\value{
Side-effect. Write csv files into the intput directory.
}
\description{
Read and write the inputs for the calculation in a tidylist.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
