% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calc.R
\name{rangeStandardize}
\alias{rangeStandardize}
\title{Standardize continuous variables}
\usage{
rangeStandardize(x)
}
\arguments{
\item{x}{Variable to standardize}
}
\description{
Function to standardize continuous variables.
}
