% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{read_description}
\alias{read_description}
\title{read the path fo the data}
\usage{
read_description(path_container)
}
\arguments{
\item{path_container}{path to container}
}
\description{
read the path fo the data
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
