% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{read_param}
\alias{read_param}
\title{Read a single parameters csv file}
\usage{
read_param(file)
}
\arguments{
\item{file}{File to read.}
}
\description{
Read a single parameters csv file.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
