% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kamila.R
\name{run_kamila}
\alias{run_kamila}
\title{Function to read the data, to execute the computations of the Kamila
clustering method and write the data and the parameters (of the inputs and
the outputs).}
\usage{
run_kamila(path, path_out = file.path(tempdir(), "out"), param_replace = NULL)
}
\arguments{
\item{path}{path of the set of parameters set to a particular output.}

\item{path_out}{path where the outputs are saved.}

\item{param_replace}{parameter which can be used to test some other parameters.}
}
\description{
This function executes the computations of the Kamila clustering
method according to the parameters which are read in a first step. In a
third step, the outputs are delivered.
}
