% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{separate_at_comma}
\alias{separate_at_comma}
\title{Separate a string of several words at commas to a vector}
\usage{
separate_at_comma(x)
}
\arguments{
\item{x}{string of words to separate}
}
\description{
Separate a string of several words at commas to a vector
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
