% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist}
\alias{tidylist}
\title{Build a tidylist}
\usage{
tidylist(...)
}
\arguments{
\item{...}{any other arguments}
}
\description{
A tidylist is a list collecting one or several tidy data frames.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
