% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist_read}
\alias{tidylist_read}
\title{Import a tidylist from a collection of tidy csv files}
\usage{
tidylist_read(path = ".", fst = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Path to the file.}

\item{fst}{If TRUE, read fst, otherwise csv.}

\item{verbose}{Print a message if TRUE.}
}
\description{
Import and read a tidylist from a collection of tidy csv files.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
