% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist_write}
\alias{tidylist_write}
\title{Function to write a list of CSV files}
\usage{
tidylist_write(x, path = ".", fst = FALSE, verbose = TRUE)
}
\description{
Function to write a list of CSV files
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
