% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\name{trace_flow}
\alias{trace_flow}
\title{Trace and visualize workflow}
\usage{
trace_flow(expr)
}
\arguments{
\item{expr}{expression for which the flow should be traced, e.g.
\code{{run_kamila(path = paths, path_out = path_out)}}.}
}
\description{
Trace and visualize workflow.
}
\author{
\href{mailto:christoph@christophsax.com}{Christoph Sax}
}
