% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{trace_this}
\alias{trace_this}
\title{Trace the inputs (\code{\link{mod_init}}) and the outputs}
\usage{
trace_this(x, at = "", mod = "")
}
\arguments{
\item{x}{Tidylist.}

\item{at}{öevel of the module ("input" or "output").}

\item{mod}{Name of the module function.}
}
\description{
Trace the inputs (\code{\link{mod_init}}) and the
outputs (\code{\link{mod_return}}).
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
