% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_computation_kamila.R
\name{wrap_computation_kamila}
\alias{wrap_computation_kamila}
\title{wrap_computation_kamila (memoised)}
\usage{
wrap_computation_kamila(tl_inp, tl_prepadata)
}
\description{
wrap_computation_kamila (memoised)
}
