% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_computation_kamila.R
\name{wrap_computation_kamila_}
\alias{wrap_computation_kamila_}
\title{Wrapper for the clusters construction.}
\usage{
wrap_computation_kamila_(tl_inp, tl_prepadata)
}
\arguments{
\item{tl_inp}{List of input data frames of which we use:
\itemize{
\item \code{PARAM_KAMILA$calc_kstar}: If TRUE, estimates the clusters. Else, takes the
parameter PARAM_KAMILA$param_kstar.
\item \code{PARAM_KAMILA$cont_var_expl}: List of continuous variables chosen as explicative
variables.
\item \code{PARAM_KAMILA$categ_var_expl}: List of categorical variables chosen as explicative
variables.
}}

\item{tl_prepadata}{List of data frames prepared in a first step.}
}
\value{
a \code{tidylist} containing the following tidylists:
\itemize{
\item \code{tl_mod_calc_kamila}
}
}
\description{
This wrapper contains all the necessary modules which allow to
construct the clusters.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
