% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_kamila.R
\name{wrap_kamila_}
\alias{wrap_kamila_}
\title{Wrapper to execute the Kamila algorithm.}
\usage{
wrap_kamila_(tl_inp_kamila)
}
\arguments{
\item{tl_inp_kamila}{tidylist of inputs}
}
\value{
a \code{tidylist} containing the following tidylists:
\itemize{
\item \code{tl_computation_kamila}
}
}
\description{
Simple function which executes the computations needed for the
Kamila algorithm.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
