% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_prepadata.R
\name{wrap_prepadata}
\alias{wrap_prepadata}
\title{Memoised copy of \code{\link{wrap_prepadata}}
Create a memoised copy of \code{\link{wrap_prepadata}}}
\usage{
wrap_prepadata(tl_inp)
}
\arguments{
\item{tl_inp}{List of input data frames.}
}
\description{
Memoised copy of \code{\link{wrap_prepadata}}
Create a memoised copy of \code{\link{wrap_prepadata}}
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
