% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_prepadata.R
\name{wrap_prepadata_}
\alias{wrap_prepadata_}
\title{Wrapper for the preparation of the data}
\usage{
wrap_prepadata_(tl_inp)
}
\arguments{
\item{tl_inp}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidylists:
\itemize{
\item \code{tl_prepa_rr}
}
}
\description{
This wrapper contains all the necessary modules which allow to
prepare the data.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
