% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{write_param}
\alias{write_param}
\title{Write parameters}
\usage{
write_param(x, file)
}
\arguments{
\item{x}{parameters tibble to write}

\item{file}{file to write to}
}
\description{
Write parameters
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
