% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{clustmeth_identifier_number}
\alias{clustmeth_identifier_number}
\title{Function to create a single identifier number}
\usage{
clustmeth_identifier_number(method_name, path)
}
\arguments{
\item{method_name}{name of the clustering method}

\item{path}{directory path}
}
\value{
identifier_number container identifier
}
\description{
Function to create a single identifier number
}
