% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_demo_data.R
\name{gen_demo_data}
\alias{gen_demo_data}
\title{Generate demo data}
\usage{
gen_demo_data(path = tempdir(), method_name = "kamila")
}
\arguments{
\item{path}{path where to store the demo data. Default: \code{tempdir()}.}

\item{method_name}{name of the clustering method.}
}
\value{
the path of the csv file
}
\description{
Generate demo data and write into a csv file
}
\examples{
gen_demo_data()
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
