% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_catcontvar.R
\name{mod_catcontvar}
\alias{mod_catcontvar}
\title{Splitting the continuous from the categorical data}
\usage{
mod_catcontvar(RR_OASI, RR_OASI_TS, RR_OASI_VS, PARAM_GLOBAL, list = NULL)
}
\arguments{
\item{RR_OASI}{a data frame containing the all the data, whose variables are:
\itemize{
\item \code{year}: Year of the pension register extract.
\item \code{age}: Age of the individual.
\item \code{age_retire}: Retirement age.
\item \code{sex}: Sex, if 1: female, if 0:male
\item \code{nat}: Nationality, if 1: Foreign, if 0:Swiss.
\item \code{resid}: Residence, if 1: Foreign, if 0:Swiss.
\item \code{benef_type1}: If 1, Old-age type of benefit (dummy)
\item \code{benef_type2}: Widow type of benefit (dummy)
\item \code{benef_type3}: Father's orphan type of benefit (dummy)
\item \code{benef_type4}: Mother's orphan type of benefit (dummy)
\item \code{benef_type5}: Twice orphan type of benefit (dummy)
\item \code{benef_type6}: Spouse's compl. type of benefit (dummy)
\item \code{benef_type7}: Father's child rent type of benefit (dummy)
\item \code{benef_type8}: Mother's child rent type of benefit (dummy)
\item \code{benef_type} : Types of benefits type of benefit (categorical)
\item \code{marital_stat1}: Divorced marital status (dummy)
\item \code{marital_stat2}: Single as reference category marital status (dummy)
\item \code{marital_stat3}: Married marital status (dummy)
\item \code{marital_stat4}: Widowed marital status (dummy)
\item \code{marital_stat}: Marital Status
\item \code{splitting}: If 1, splitting of the revenues, 0 otherwise.
\item \code{capping}:  If 1, the pension is capped, 0 otherwise.
\item \code{contrib_m_ind}: total number of OASI contribution months per individual.
\item \code{contrib_y_ageclass}: total number of contribution years per age group.
\item \code{bonus_m_edu}: number of months paid with a bonus for educative tasks.
\item \code{bonus_m_assist}: number of months paid with a bonus for assistance/care
tasks.
}}

\item{RR_OASI_TS}{a training dataset containing x\% of the data.}

\item{RR_OASI_VS}{a validation dataset containing (100 - x)\% of the data.}

\item{PARAM_GLOBAL}{a data frame containing the parameters. We use the
following:
\itemize{
\item \code{categ_var}: Chosen categorical variables
\item \code{cont_var}: Chosen continuous variables
}}

\item{list}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidy data frames:
\itemize{
\item \code{CATEG_DF}: Contains only categorical variables (factors).
\item \code{CONT_DF} : Contains only continuous variables (numeric).
\item \code{CATEG_DF_TS}: Contains only categorical variables (factors),
training set.
\item \code{CONT_DF_TS} : Contains only continuous variables (numeric),
training set.
\item \code{CATEG_DF_VS}: Contains only categorical variables (factors),
validation set.
\item \code{CONT_DF_VS} : Contains only continuous variables (numeric),
validation set.
}
}
\description{
Splits the continuous from the categorical data, in order to use
a clustering method.
}
\references{
\href{https://www.geeksforgeeks.org/the-validation-set-approach-in-r-programming/}{www.geeksforgeeks.org}
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
