% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{mod_init}
\alias{mod_init}
\title{mod_init() pour les fonctions mod_}
\usage{
mod_init(mod.function = NULL)
}
\arguments{
\item{mod.function}{module}
}
\description{
Vérifie si tous les inputs sont des tibbles.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
