% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_out.R
\name{mod_out_kamila}
\alias{mod_out_kamila}
\title{Output Kamila}
\usage{
mod_out_kamila(path, path_out, tl_inp_kamila, tl_out_kamila)
}
\arguments{
\item{path}{Vector containing container path.}

\item{path_out}{Vector containing output path.}

\item{tl_inp_kamila}{List containing input data Kamila.}

\item{tl_out_kamila}{List containing output data Kamila.}
}
\value{
\code{path_out_identifier}: Path to output container.
}
\description{
Saves the Kamila output
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
