% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{mod_return}
\alias{mod_return}
\title{mod_return() pour les fonctions mod_}
\usage{
mod_return(...)
}
\description{
Vérifie si tous les outputs sont des tibbles.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
