% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{path_data}
\alias{path_data}
\title{Read the data from the path in PARAM_GLOBAL}
\usage{
path_data(path)
}
\arguments{
\item{path}{directory path}
}
\value{
path_data
}
\description{
Read the data from the path in PARAM_GLOBAL
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
