% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{prepare_input}
\alias{prepare_input}
\title{Preparation of the inputs for the calculation}
\usage{
prepare_input(path, path_out = file.path(dirname(path)), overwrite = FALSE)
}
\arguments{
\item{path}{path to container}
}
\value{
nothing. Output (the input data) is stored in the assigned location
}
\description{
Reading, preparing and saving the inputs.
}
\author{
\href{mailto:layalchristine.lettry@unifr.ch}{Layal Christine Lettry}
}
