% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{read_path_data}
\alias{read_path_data}
\title{Read the path fo the data}
\usage{
read_path_data(path_container)
}
\arguments{
\item{path_container}{Path to container.}
}
\description{
Read the path fo the data
}
