% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_description.R
\name{read_path_data}
\alias{read_path_data}
\title{read the path fo the data}
\usage{
read_path_data(
  path_container =
    "/Users/christoph/local/bsv/container_abr18def_rr18_statpop2017_estv3_va20004_real_stafasgo"
)
}
\description{
read the path fo the data
}
