% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrclust-package.R
\docType{package}
\name{rrclust-package}
\alias{rrclust}
\alias{rrclust-package}
\title{rrclust: Clustering the Swiss Pension Register}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package allows to cluster the Swiss Pension Register by using the Kamila clustering method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asam-group/rrclust}
}

}
\keyword{internal}
