% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_kamila.R
\name{run_kamila}
\alias{run_kamila}
\title{Run the Kamila estimation}
\usage{
run_kamila(path, path_out = file.path(tempdir(), "out"), param_replace = NULL)
}
\arguments{
\item{path}{Path of the set of parameters set to a particular output.}

\item{path_out}{Path where the outputs are saved.}

\item{param_replace}{Parameter which can be used to test some other
parameters.}
}
\description{
This function executes the computations of the Kamila clustering method
according to the parameters which are read in a first step. Eventually, the
outputs are written into csv or excel files.
}
