% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{separate_at_comma}
\alias{separate_at_comma}
\title{Separate a string of several words at commas to a vector}
\usage{
separate_at_comma(x)
}
\arguments{
\item{x}{String of words to split.}
}
\description{
Split a character string of several words at commas to a vector.
}
\examples{
x <- "sex, nat, resid, benef_type1, benef_type2"
separate_at_comma(x)
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
