% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist}
\alias{tidylist}
\title{Creator function for tidylists}
\usage{
tidylist(...)
}
\arguments{
\item{...}{any other arguments}
}
\description{
Creator function for tidylists
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
