% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist_ensure}
\alias{tidylist_ensure}
\title{Ensure a tidylist contains only 'tidy' dfs}
\usage{
tidylist_ensure(x)
}
\description{
Ensure a tidylist contains only 'tidy' dfs
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
