% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist_read}
\alias{tidylist_read}
\title{Import a Tidylist from a Collection of tidy CSV files}
\usage{
tidylist_read(path = ".", fst = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{character, containg the path(s) to a directory or to one or several CSV files.}
}
\description{
Import a Tidylist from a Collection of tidy CSV files
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
