% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidylist.R
\name{tidylist_write}
\alias{tidylist_write}
\title{Write a list of CSV files}
\usage{
tidylist_write(x, path = ".", fst = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{Object to write.}

\item{path}{Path to file.}

\item{fst}{If TRUE, write fst, otherwise csv.}

\item{verbose}{Print a message if TRUE.}
}
\description{
Write a list of CSV files.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
