% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infrastructure.R
\name{trace_this}
\alias{trace_this}
\title{trace_this()}
\usage{
trace_this(x, at = "", mod = "")
}
\description{
Retourne "do not use \code{mod_init()} in functions without arguments." si aucun argument n'est donné dans la fonction.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
