% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_kamila.R
\name{wrap_kamila}
\alias{wrap_kamila}
\title{Memoised copy of \code{\link{wrap_kamila}}
Create a memoised copy of \code{\link{wrap_kamila}}}
\usage{
wrap_kamila(tl_inp_kamila)
}
\arguments{
\item{tl_inp_kamila}{tidylist of inputs}
}
\description{
Memoised copy of \code{\link{wrap_kamila}}
Create a memoised copy of \code{\link{wrap_kamila}}
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
