% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_kamila.R
\name{wrap_kamila_}
\alias{wrap_kamila_}
\title{Wrapper to execute the Kamila algorithm.}
\usage{
wrap_kamila_(tl_inp_kamila)
}
\arguments{
\item{tl_inp_kamila}{tidylist of inputs}
}
\value{
\code{tl_computation_kamila} with the following data frames:
\itemize{
\item \code{PLOTDATKAM} Data frame containing the clusters factor and the other
variables.
\item \code{KM_RES_FINAL} Data frame containing the resulting parameters of the
clustering.
\item \code{CONTVARS} Data frame containing the continuous standardised variables.
\item \code{FULL_CONT_DF} Data frame containing the continuous variables used for
the estimation.
\item \code{FULL_CATEG_DF} Data frame containing the categorical variables used for
the estimation.
}
}
\description{
Simple function which executes the computations needed for the
Kamila algorithm.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
