% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_prepadata.R
\name{wrap_prepadata_}
\alias{wrap_prepadata_}
\title{Wrapper for the preparation of the data}
\usage{
wrap_prepadata_(tl_inp)
}
\arguments{
\item{tl_inp}{List of input data frames.}
}
\value{
a \code{tidylist} containing the following tidylists:
\itemize{
\item \code{tl_mod_catcontvar}: A \code{tidylist} containing the following data frames:
\itemize{
\item \code{CATEG_DF}: Contains only categorical variables (factors).
\item \code{CONT_DF} : Contains only continuous variables (numeric).
\item \code{CATEG_DF_TS}: Contains only categorical variables (factors),
training set.
\item \code{CONT_DF_TS} : Contains only continuous variables (numeric),
training set.
\item \code{CATEG_DF_VS}: Contains only categorical variables (factors),
validation set.
\item \code{CONT_DF_VS} : Contains only continuous variables (numeric),
validation set.
}
}
}
\description{
This wrapper contains all the necessary modules which allow to
prepare the data.
}
\author{
\href{mailto:layal.lettry@gmail.com}{Layal Christine Lettry}
}
