% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_clust.R
\name{write_param}
\alias{write_param}
\title{Write parameters}
\usage{
write_param(x, file)
}
\arguments{
\item{x}{Parameters tibble to write.}

\item{file}{File path to write to.}
}
\description{
Write parameters into a file.
}
\author{
\href{mailto:christoph@cynkra.com}{Christoph Sax}
}
